Change Clean to go back to real original material
I found a way to adjust your addon so when the scene is cleaned they return to the materials they had before. That method is not elegant at all though, it relys soley on the material names and the materials that do still exist in the .blend to figure out which material to give which object when cleaning. This means it changes the names of materials, which probably breakes some stuff, though I have not discorved that myself when using it so far. The "XX_Original" materials are here not tagged as "Fake user", because they are actually not needed anymore since the really original Material is being tagged as fake user and the reaplied to the objects it has previously been applied to.
Also my formatter changed a bunch of files in meaningless ways.

+++ Addon/Operators/clean.py
@@ -1,12 +1,16 @@
-import bpy, os, shutil
-from .. Utility import utility, matcache
+import bpy
+import os
+import shutil
+from ..Utility import utility, matcache
+
 
 class TLM_CleanLightmaps(bpy.types.Operator):
     """Cleans the lightmaps"""
+
     bl_idname = "tlm.clean_lightmaps"
     bl_label = "Clean Lightmaps"
     bl_description = "Clean Lightmaps"
-    bl_options = {'REGISTER', 'UNDO'}
+    bl_options = {"REGISTER", "UNDO"}
 
     def execute(self, context):
 
@@ -21,21 +25,28 @@ class TLM_CleanLightmaps(bpy.types.Operator):
             if obj.TLM_ObjectProperties.tlm_mesh_lightmap_use:
                 obj["Lightmap"] = ""
                 if hasattr(obj, "Lightmap"):
-                    del(obj["Lightmap"])
+                    del obj["Lightmap"]
 
         if scene.TLM_SceneProperties.tlm_clean_option == "Selection":
             filepath = bpy.data.filepath
-            dirpath = os.path.join(os.path.dirname(bpy.data.filepath), scene.TLM_SceneProperties.tlm_lightmap_savedir)
+            dirpath = os.path.join(
+                os.path.dirname(bpy.data.filepath),
+                scene.TLM_SceneProperties.tlm_lightmap_savedir,
+            )
             if os.path.isdir(dirpath):
                 list = os.listdir(dirpath)
                 for file in list:
                     for obj in selection:
-                        if file.startswith(obj.name[:4]):
+                        # As far as I understand the code the second condition shouldn't be necessary but if it isn't there
+                        # sometimes an error is thrown because the file doesn't exist.
+                        if file.startswith(obj.name[:4]) and os.path.exists(
+                                os.path.join(dirpath, file)):
                             print(file)
+                            # Why? Both cases do the same thing
                             if file.endswith(".pfm"):
-                                os.remove(os.path.join(dirpath,file))
+                                os.remove(os.path.join(dirpath, file))
                             if file.endswith(".hdr"):
-                                os.remove(os.path.join(dirpath,file))
+                                os.remove(os.path.join(dirpath, file))
 
             for obj in selection:
                 for slot in obj.material_slots:
@@ -43,25 +54,31 @@ class TLM_CleanLightmaps(bpy.types.Operator):
 
         elif scene.TLM_SceneProperties.tlm_clean_option == "Clean cache":
             filepath = bpy.data.filepath
-            dirpath = os.path.join(os.path.dirname(bpy.data.filepath), scene.TLM_SceneProperties.tlm_lightmap_savedir)
+            dirpath = os.path.join(
+                os.path.dirname(bpy.data.filepath),
+                scene.TLM_SceneProperties.tlm_lightmap_savedir,
+            )
             if os.path.isdir(dirpath):
                 list = os.listdir(dirpath)
                 for file in list:
                     if file.endswith(".pfm"):
-                        os.remove(os.path.join(dirpath,file))
+                        os.remove(os.path.join(dirpath, file))
                     if file.endswith("denoised.hdr"):
-                        os.remove(os.path.join(dirpath,file))
+                        os.remove(os.path.join(dirpath, file))
                     if file.endswith("finalized.hdr"):
-                        os.remove(os.path.join(dirpath,file))
+                        os.remove(os.path.join(dirpath, file))
 
             for image in bpy.data.images:
                 if image.name.endswith("_baked"):
                     image.save()
 
-        else: #Clean and Restore
+        else:  # Clean and Restore
             print("Clean and restore...")
             filepath = bpy.data.filepath
-            dirpath = os.path.join(os.path.dirname(bpy.data.filepath), scene.TLM_SceneProperties.tlm_lightmap_savedir)
+            dirpath = os.path.join(
+                os.path.dirname(bpy.data.filepath),
+                scene.TLM_SceneProperties.tlm_lightmap_savedir,
+            )
             if os.path.isdir(dirpath):
                 for file in os.listdir(dirpath):
                     os.remove(os.path.join(dirpath + "/" + file))
@@ -77,14 +94,14 @@ class TLM_CleanLightmaps(bpy.types.Operator):
                     bpy.data.images.remove(image, do_unlink=True)
 
         for mat in bpy.data.materials:
-        #     if mat.name.endswith('_Original'):
-        #         bpy.data.materials.remove(mat, do_unlink=True)
-        #     if mat.name.endswith('.temp'):
-        #         bpy.data.materials.remove(mat, do_unlink=True)
-            if mat.name.endswith('_temp'):
+            #     if mat.name.endswith('_Original'):
+            #         bpy.data.materials.remove(mat, do_unlink=True)
+            #     if mat.name.endswith('.temp'):
+            #         bpy.data.materials.remove(mat, do_unlink=True)
+            if mat.name.endswith("_temp"):
                 bpy.data.materials.remove(mat, do_unlink=True)
 
         for mat in bpy.data.materials:
             mat.update_tag()
 
-        return{'FINISHED'}
+        return {"FINISHED"}
